python/requirements.txt
discord_webhook==1.0.0
notion_client==2.0.0
python-dotenv==0.21.1
